Updated README.md with working example resolvers
The current example for Bare Workflow resolvers does not work. This is because the additional resolvers are not included with the `standard-version-expo` package. They are only available from this repo, which needs to be specified when requiring the updater.

Also, the `android/app/build.gradle` updater` should not use `require.resolve` as we resolve it when calling it with the `sdkVersion` as a parameter.